requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
numpy==2.3.1
pymongo==4.13.2
torch==2.7.1
sentence-transformers==5.0.0